Make it clearer where the lines of a file are being read
This is to group together calls to a function that reads a line from
file with calls to a function that gets the line, so it is easier to
reason about the code. On top of that, reading the file also happened
in a function that was running the handlers of various pieces of an
annotated patch and the handlers relied on that. It was a bit
confusing.

This commit also introduces the possibility to unread some lines,
which became needed after the above changes. This is because the
format of a patch sometimes does not give me enough context to tell
that a different part of an annotated patch has begun. I can learn
about it after reading the first line of the new part. So to keep the
part handlers simple, I unread the line and let the handler read it
again.