Collect stats while parsing the annotated patch
There is a hack for the binary patches - even if we parsed the actual
binary patch, we can't always tell what was the size of the binary
file before and after applying the patch. This is the case especially
when modifying a part of the binary file. We need to take this
information from listing in the annotated patch. The problem is that
the paths there can be truncated, so we need to compute the relevancy
of the line in the listing to the changed binary file.

There could be still a bit more heuristics applied, but I will stick
with what I got so far for know.