Capture output of commands in test script
Instead of capturing an output only of the split.pl script, capture
the output from all the git commands and some others too. The solution
is a generalized a bit, but still quite verbose.

This also adds a function for printing the status, so I can change the
structure of the code a bit. Actually I could lower the indentation
for one level for a large block, but will do it in the following
commit.