Split the splitter
Put every package in the script into a separate file inside the Kgps
directory (and thus this puts them into a Kgps namespace). This
currently makes the script harder to install or use, because you need
to either put all the new files in a place that perl can find the
modules or use `perl -Idir/with/the/files`.

Generation of the standalone script should follow soon.