Fix a bug when getting a context from chunk without it
Hardcoding @{foo}[0 .. 2] caused an autovivification of the array
items, if the array has fewer than 3 of them. Since the autovivified
items were undefs, the script was bailing out at some point, because
it was expecting them to be objects on which some methods could be
called.

Now instead of hardcoding the range, we take an actual length of the
array into account, so no autovivification happen.