Add tests for handling some edge cases in handling empty files
There are four tests. One that appends lines to a file, one that
prepends lines to a file, one that deletes lines from the beginning of
the file and one that deletes lines from the end of a file.

The tests that adds lines to a file start with an empty file. The
tests that remove lines from a file end up with an empty file. So this
helps us trigger the edge cases.