Fix two issues related to special 0 line no in marker
Line no in location marker being zero is a special case - it means
that the file has no lines. If the file has no lines then the line
count (the number after comma in location marker) also needs to be
zero. So at the same time either both line no and line count need to
be zero or both need to be greater than zero.

This needs to be taken into account when doing some maths on the
marker. For example, incrementing a line no for younger sections from
0 should result in 2, because the older section will have 1. Or when
line count after subtraction is zero, line no should also become zero.

So off-by-one errors were just waiting to show their ugly faces and so
they did. This commit drops the off-by-one corrections done
postprocessing as it was not enough and the corrected maths in
adapting markers are already taking care of possible transitions
between zero and non-zero numbers in markers.

Probably similar situation happens in the code that handles overlapped
sections, but to fix possible bugs there I need to have a test case
first.