Sort some uses
Maybe not that useful from pragmas (they are hardcoded for now), but
certainly for other packages. They are pulled from the parsed modules
and put at the top of the generated script, and their order depended
on which module was parsed first. It would be nice to have them in a
predictable order, so sort them.