Add an initial form of time increments handling
This adds a DATE_INC clause that can appear only before the first
section. Two modes of increments are supported - FROM and UPTO. FROM
takes the date from the patch, and for each section the clause
increments the date by the time from clause. UPTO is similar, but the
patch date is by default treated like the latest date for a patch, so
in this mode the clause goes through sections in reverse order and
decrements the date by the time from clause.

If the section has a DATE clause, then this date is used for the patch
directly and this date serves as a base of furter
increments/decrements of the patch date.