new test: Add a simple test for handling rename diff header
It's a bit artificial, because it is not possible to generate a patch
that only contains a rename, without code changes in the file. This is
because we don't have a way to tell the script in the annotated patch,
that a certain section should do only a rename, without any other
changes. Trying to do it by putting SECTION clause in the beginning of
the file in a way that it will contain only the context of the diff
will result in bail out with a message about a useless section in the
diff.