Add a script for easy test generation
Creating a test case involves a bunch of repetitive steps, so let's
automate that a bit. The script sets up the proper directory structure
for the testcase, prepares a temporary git repo for the test case and
puts proper patches in proper places after the manual part of an
actual creation of the test case body. After the script is done, what
is left is to edit the test.patch.